# Adding mix_test_watch
Moving forward, we'll be writing tests for the Elixir code we create.
Being the good developers that we are, we should test this code.

To make testing a more integrated process, we'll add the
`mix_text_watch` dependency, which lets us run `mix test.watch` command
to continuously watch our project for changes and re-run our test suite.