# Layouts and Containers
Now that we have our basic React functionality set up and our static
assets being served, it's time to start migrating the React components
from the Meteor Todos application into our new Phoenix application.

We'll start this process by changing our `app.html.eex` file to use the
expected `"app"` ID on its container div.

    <div id="app"></div>

Next, we can update our `app.js` file, removing our `HelloWorld`
component, and replacing it with the setup found in the Todos
application. We need to be sure to remove the `Meteor.startup`
reference, as we won't be using Meteor:

    import { render } from "react-dom";
    import { renderRoutes } from "./routes.jsx";

    render(renderRoutes(), document.getElementById("app"));

Now we port over the `routes.jsx` file. We'll put this directly into our
`web/static/js` folder, next to our `app.js` file.

We'll keep things simple at first by only defining routes for the
`AppContainer` and the `NotFoundPage`.

    import AppContainer from './containers/AppContainer.jsx';
    import NotFoundPage from './pages/NotFoundPage.jsx';

    export const renderRoutes = () => (
      <Router history={browserHistory}>
        <Route path="/" component={AppContainer}>
          <Route path="*" component={NotFoundPage}/>
        </Route>
      </Router>
    );

The `AppContainer` in the Meteor application defines a reactive
container around an `App` component. This is very Meteor-specific, so
we'll gut this for now and replace it with a simple container that sets
up our initial application state and passes it down to the `App`
component.

Next comes the process of migrating the `App` component and all of its
children components (`UserMenu`, `ListList`, `ConnectionNotification`,
etc...).

This migration is fairly painless. We just need to be sure to
remove references to Meteor-specific functionality. We'll replace the
functionality we remove soon.

After all of these changes, we're greeted with a beautifully styped
loading screen when we refresh our application.