# Fixing Routes
Unfortunately, our previous changes to `app.html.eex` broke our page
controller test. The test was expecting the result of getting the root
URL of the application to contain the string `"Welcome to Phoenix"`.

We removed that portion of the template, which caused the test to fail.
A better test is to assert that the response contains our `"app"` div:

    assert html_response(conn, 200) =~ "<div id=\"app\"></div>"

This brings our test suite back into the green.

While using our application, we also noticed another route-related bug.
When navigating to a page using React Router, like `/signin`, and
refreshing the page, the server would return a `404` response,
complaining that it can't find a route associated with `/signin`.

Because we're building a single page application, all page requests
should render our main application template. We can write a quick test
for this:

    test "GET /signin", %{conn: conn} do
      conn = get conn, "/signin"
      assert html_response(conn, 200) =~ "<div id=\"app\"></div>"
    end

And we can make a quick change in our router to use our `PageController`
for all requests:

    get "/*page", PageController, :index

With that change, our tests go back to green, and our application
functions as expected.