# Create Users Table
Let's focus on adding the notions of users and authorization to our
Todos application. The first thing we'll need is to create a database
table to hold our users, and a corresponding schema.

Thankfully, Phoenix comes with many
[generators](http://www.phoenixframework.org/docs/mix-tasks) that make
generating migrations and models very easy.

To generate our users migration, we'll run the following `mix` command:

    mix phoenix.gen.model User users email:string encrypted_password:string

We'll modify the migration file the generator created for us and add
`NOT NULL` restrictions on both the `email` and `encrypted_password`
fields:

    add :email, :string, null: false
    add :encrypted_password, :string, null: false

We'll also add an index on the `email` field for faster queries:

    create unique_index(:users, [:email])

Great! Now we can run that migration with the `mix ecto.migrate`
command.