# Creating the Users Model
Now that we're created our users table, we need to create a
corresponding `User` model. Phoenix actually did most of the heavy
lifting for us when we ran the `mix phoenix.gen.model` command.

If we look in `/web/models`, we'll find a `user.ex` file that holds our
new `User` model. While the defaults generated for us are very good,
we'll need to tweak a couple things.

In addition to the `:email` and `:encrypted_password` fields, we'll also
need a [virtual](https://hexdocs.pm/ecto/Ecto.Schema.html#field/3)
`:password` field.

    field :password, :string, virtual: true

`:password` is virtual because it will be required by our `changeset`
function, but will not be stored in the datbase.

Speaking of required fields, we'll need to update our `@required_fields`
and `@optional_fields`
[attributes](http://elixir-lang.org/getting-started/module-attributes.html)
to reflect the changes we've made:

    @required_fields ~w(email password)
    @optional_fields ~w(encrypted_password)

These changes to `@required_fields` broke our auto-generated tests
against the `User` model. We'll need to update the `@valid_attrs`
attribute in `test/models/user_test.ex` and replace
`:encrypted_password` with `:password`:

  @valid_attrs %{email: "user@example.com", password: "password"}

And with that, our tests flip back to green!