# Additional Validation
While the default required/optional field validation is a good start, we
know that we'll need additional validation on our `User` models.

For example, we don't want to accept email addresses without the `"@"`
symbol. We can write a test for this in our `UserTest` module:

    test "changeset with invalid email" do
      changeset = User.changeset(%User{}, %{
        email: "no_at_symbol",
        password: "password"
      })
      refute changeset.valid?
    end

Initially this test fails, but we can quickly make it pass by adding
basic regex validation to the `:email` field in our `User.changeset`
function:

    |> validate_format(:email, ~r/@/)

We can repeat this process for all of the additional validation we need,
like checking password length, and asserting email uniqueness.