# Hashing Our Password
You might have noticed that we had to manually set values for the
`encrypted_password` field for our `"changeset with non-unique email"`
test to run. This was to prevent the database from complaining about a
non-null constraint violation.

Let's remove those lines from our test and generate the password hash
ourselves!

We'll use the `comeonin` package to hash our passwords, so we'll add it
as a dependency and an application in `mix.exs`:

    def application do
      [...
       applications: [..., :comeonin]]
    end

    defp deps do
      [...
       {:comeonin, "~> 2.0"}]
    end

Now we can write a private method that will update the our
`:encrypted_password` field on our `User` model if its given a valid
changeset that's updating the value of `:password`:

    defp put_encrypted_password(changeset = %Ecto.Changeset{
      valid?: true,
      changes: %{password: password}
    }) do
      changeset
      |> put_change(:encrypted_password, Comeonin.Bcrypt.hashpwsalt(password))
    end

We'll use pattern matching to handle the cases where a changeset is
either invalid, or not updating the `:password` field:

    defp put_encrypted_password(changeset), do: changeset

Isn't that pretty? And with that, our tests are passing once again.