# Enter Guardian
Now we're getting to the meat of our authentication system. We have our
`User` model set up, but now we need to associate users with active
sessions.

This is where [Guardian](https://github.com/ueberauth/guardian) comes
in. Guardian is an authentication framework that leverages JWT and plays
nicely with Phoenix Channels.

To use Guardian, we'll first add it as a depenency to our application:

    {:guardian, "~> 0.12.0"}

Next, we need to do some configuring:

    config :guardian, Guardian,
      allowed_algos: ["HS512"], # optional
      verify_module: Guardian.JWT,  # optional
      issuer: "PhoenixTodos",
      ttl: { 30, :days },
      verify_issuer: true, # optional
      secret_key: System.get_env("GUARDIAN_SECRET_KEY"),
      serializer: PhoenixTodos.GuardianSerializer

You'll notice that I'm pulling my `secret_key` from my system's
environment variables. It's a bad idea to keep them in version control.

I also specified a `serializer` module. This is Guardian's bridge into
your system. It acts as a translation layer between Guardian's JWT and
your `User` model.

Because it's unique to our system, we'll need to build the
`PhoenixTodos.GuardianSerielizer` ourselves.

We'll need to define two fuctions. The first, `for_token` translates a
`User` model into a token string. An invalid `User` should return an
`:error`:

    test "generates token for valid user", %{user: user} do
      assert {:ok, _} = GuardianSerializer.for_token(user)
    end

    test "generates error for invalid user", %{} do
      assert {:error, "Invalid user"} = GuardianSerializer.for_token(%{})
    end

Thanks to Elixir's pattern matching, `for_token` is a very simple
function:

    def for_token(%User{id: id}), do: {:ok, "User:#{id}"}
    def for_token(_), do: {:error, "Invalid user"}

Similarly, we need to define a `from_token` function, which takes a
token string and returns the corresponding `User` model:

    test "finds user from valid token", %{user: user} do
      {:ok, token} = GuardianSerializer.for_token(user)
      assert {:ok, _} = GuardianSerializer.from_token(token)
    end

    test "doesn't find user from invalid token", %{} do
      assert {:error, "Invalid user"} = GuardianSerializer.from_token("bad")
    end

To implement this, we'll pull the `User` id out of the token string, and
look it up in the database:

    def from_token("User:" <> id), do: {:ok, Repo.get(User,
  String.to_integer(id))}
    def from_token(_), do: {:error, "Invalid user"}

Now that we've finished all of that, we're in a position to wire up the
rest of our authentication system!