# Join Route and Controller
Now that all of the pieces are in place, we can finally wire up a route
to our join functionality.

To do this, we'll create an `"/api/users"` route that sends `POST`
requests to the `UserController.create` function:

    post "/users", UserController, :create

We expect the user's `email` and `password` to be sent as parameters to
this endpoint. `UserController.create` takes those `params`, passes them
into our `User.changeset`, and then attempts to `insert` the resulting
`User` into the database:

    User.changeset(%User{}, params)
    |> Repo.insert

If the `insert` fails, we return the `changeset` errors to the client:

    conn
    |> put_status(:unprocessable_entity)
    |> render(PhoenixTodos.ApiView, "error.json", error: changeset)

Otherwise, we'll use Guardian to sign the new user's JWT and return the
`jwt` and `user` objects down to the client:

    {:ok, jwt, _full_claims} = Guardian.encode_and_sign(user, :token)
    conn
    |> put_status(:created)
    |> render(PhoenixTodos.ApiView, "data.json", data: %{jwt: jwt, user: user})

Now all a user needs to do to join the Todos application is send a
`POST` request to `/api/users` with their email and password. In turn,
they'll receive their JWT which they can send along with any subsequent
requests to verify who they are.