# Front-end Join
Now that our back-end is shored up, we can turn our attention to the
front-end side of the sign-up functionality.

The first thing we need to do is add the `"join"` route to our router:

    <Route path="join" component={AuthPageJoin}/>

We can copy the majority of the `AuthPageJoin` component from the Meteor
Todos project.

One small change we need to make is to rename all references to
`confirm` to `password_confirm` to match what our `User.changeset`
expects.

We'll also need to refactor how we create the user's account. Instead of
using Meteor's `Accounts` system, we'll need to manually make a `POST`
request to `"/api/users"`, passing the user provided `email`,
`password`, and `password_confirm` fields.

    fetch("/api/users", {
      method: "post",
      headers: {
        "Accept": "application/json",
        "Content-Type": "application/json",
      },
      body: JSON.stringify({
        user: {
          email,
          password,
          password_confirm
        }
      })
    })

If we receive any `errors` from the server, we can drop them directly
into the `errors` field of our component's state:

    let errors = json.errors.reduce((errors, error) => {
      return Object.assign(errors, error);
    });
    this.setState({errors});

Otherwise, if everything went well we'll recieve the newly signed up
user's JWT and user object in the `json.jwt` and `json.user` fields.