# Sign In Route and Controller
Now that users have the ability to join our application, how will they
sign into their accounts?

We'll start implementing "Sign In" functionality by adding a new route
to our Phoenix application:

    post "/sessions", SessionController, :create

When a user sends a `POST` request to `/sessions`, we'll route them to
the `create` function in our `SessionController` module. This function
will attempt to sign the user in with the credentials they provide.

At a high level, the `create` function will be fairly straight-forward.
We want to look up the user based on the `email` they gave, check if the
`password` they gave matches what we have on file, and then handle the
response:

    def create(conn, %{"email" => email, "password" => password}) do
      user = get_user(email)
      user
      |> check_password(password)
      |> handle_check_password(conn, user)
    end

If `get_user` returns `nil`, we couldn't find the user based on the
email address they provided. In that case, we'll return `false` from
`check_password` to return our error response:

    defp check_password(nil, _password), do: false

Otherwise, we'll use `Comeonin` to compare the hashed password we have
saved in `encrypted_password` with the hash of the password the user
provided:

    defp check_password(user, password) do
      Comeonin.Bcrypt.checkpw(password, user.encrypted_password)
    end

If all goes well, we'll return a `jwt` and the `user` object for the
now-authenticated user:

    render(PhoenixTodos.ApiView, "data.json", data: %{jwt: jwt, user: user})

We can test this sign-in route/controller combination just like we've
tested our join functionality.