# Sign In Front-end
Finally, we can continue the same pattern we've been following and
implement our sign-in functionality.

We'll start by copying over the `SignInPage` component from our Meteor
application. Next, we'll make three new actions: `SIGN_IN_REQUEST`,
`SIGN_IN_SUCCESS`, and `SIGN_IN_FAILURE`.

In addition to our actions, we'll make an asynchronous action creator
that sends a `POST` request to `/api/sessions` to initiate a sign-in.

The reducers for our new actions will be identical to our sign-up
reducers, so we'll save some typing and re-use them:

    case SIGN_IN_SUCCESS:
    case SIGN_UP_SUCCESS:
      return Object.assign({}, state, {
        user: action.user,
        jwt: action.jwt
      });

Lastly, we can replace the call to `Meteor.loginWithPassword` with a
call to our `signIn` helper. If this call is successful, we'll redirect
to `/`:

    this.state.signIn(email, password)
      .then((success) => {
        if (success) {
          this.context.router.push('/');
        }
      });

Otherwise, we'll render any errors we find in `this.props.errors`:

    const errors = (this.props.errors || []).reduce((errors, error) => {
      return Object.assign(errors, error);
    }, {});

And with those changes, a user can now sign up, log out, and sign into
our application!