# Sign Out Route and Controller
The final piece of our authorization trifecta is the ability for users
to sign out once they've successfully joined or signed into the
application.

To implement sign out functionality, we'll want to create a route that
destroys a user's session when its called by an authenticated user:

    delete "/sessions", SessionController, :delete

This new route points to `SessionController.delete`. This function
doesn't exist yet, so let's create it:

    def delete(conn, _) do
      conn
      |> revoke_claims
      |> render(PhoenixTodos.ApiView, "data.json", data: %{})
    end

`revoke_claims` will be a private function that simply looks up the
current user's token and [claims](https://jwt.io/introduction/#payload),
and then revokes them:

    {:ok, claims} = Guardian.Plug.claims(conn)
    Guardian.Plug.current_token(conn)
    |> Guardian.revoke!(claims)

In implementing this feature, we cleaned up our `SessionControllerTest`
module a bit. We added a `create_user` function, which creates a user
with a given email address and password, and a `create_session` function
that logs that user in.

Using those functions we can create a user's session, and then construct
a `DELETE` request with the user's JWT (`session_response["jwt"]`) in
the `"authorization"` header. If this request is successful, we've
successfully deleted the user's session.