# Enter Redux
Now that we're receiving the newly signed up user and JWT from the
server, we'll need some way of updating our client-side application
state.

Previously, we were doing this with [Meteor's
`createContainer`](https://guide.meteor.com/react.html#using-createContainer)
function. Since we're not using Meteor, this is no longer an option.

Instead, we'll opt for using [Redux](http://redux.js.org/).

The first thing we need to do to get going with Redux is to install our
NPM dependencies:

    npm install --save redux react-redux

Now we need to think about the "shape" of our application's state.
Thankfully, we don't have to think too hard, as this has mostly been
decided for us in the old `AppContainer` component. We'll define this
"state shape" in a new file, `/web/static/js/reducers.js`:

    const initialState = {
        user: undefined,
        jwt: undefined,
        loading: false,
        connected: true,
        menuOpen: false,
        lists: []
    };

Notice that we slipped in a new field: `jwt`. We'll use this field to
hold the JWT we get as a response from our server when signing in or
signing up.

Now we need to define our
[reducer](http://redux.js.org/docs/basics/Reducers.html):

    export default (state = initialState, action) => {
        switch(action.type) {
            default:
                return state;
        }
    }

Since we don't have any
[actions](http://redux.js.org/docs/basics/Actions.html) defined yet, our
reducer is as simple as it gets.

Now that we have our reducer defined, we can create our
[store](http://redux.js.org/docs/basics/Store.html) in `app.js`:

    const store = createStore(reducers);

We'll pass our store into our `renderRoutes` function so we can wrap our
router in a `<Provider>` component:

    <Provider store={store}>
      <Router ...>
        ...
      </Router>
    </Provider>

And lastly, we'll use `subscribe` to trigger a re-render any time our
`store` changes.

Now that our Redux store is all wired up, we can pull the state
initialization out of our `AppContainer` component and connect it to our
Redux store:

const AppContainer = connect(state => state)(App);

Now our application state is passed from our store, into our
`AppContainer`, and down into the `App` component.