# Sign Up Actions
Now that we're using Redux, we'll want to create actions that represent
out sign-up process. We'll start with three actions: `SIGN_UP_REQUEST`,
`SIGN_UP_SUCCESS`, and `SIGN_UP_FAILURE`:

    export function signUpRequest() {
      return { type: SIGN_UP_REQUEST };
    }

    export function signUpSuccess(user, jwt) {
      return { type: SIGN_UP_SUCCESS, user, jwt };
    }

    export function signUpFailure(errors) {
      return { type: SIGN_UP_FAILURE, errors };
    }

We'll also pull the `fetch` call that we're using to hit our
`/api/users` endpoint into a helper method called `signUp`. `signUp`
will dispatch a `SIGN_UP_REQUEST` action, followed by either a
`SIGN_UP_SUCCESS` or `SIGN_UP_FAILURE`, depending on the result of our
`fetch`.

Because we're using asynchronous actions, we'll need to pull in the
[Redux Thunk](https://github.com/gaearon/redux-thunk) middleware:

    npm install --save redux-thunk

And then wire it into our store:

    const store = createStore(
      reducers,
      applyMiddleware(thunkMiddleware)
    );

Now that our actions are defined, we'll need to create a matching set of
reducers. The `SIGN_UP_REQUEST` reducer does nothing:

    case SIGN_UP_REQUEST:
      return state;

The `SIGN_UP_SUCCESS` reducer stores the returned `user` and `jwt`
object in our application state:

    case SIGN_UP_SUCCESS:
      return Object.assign({}, state, {
        user: action.user,
        jwt: action.jwt
      });

And the `SIGN_UP_FAILURE` reducer stores the returned `errors` (you'll
also notice that we added an `errors` field to our `initialState`):

    case SIGN_UP_FAILURE:
      return Object.assign({}, state, {
        errors: action.errors
      });

Great. Now we can wrap our `JoinPage` component in a Redux `connect`
wrapper and pull in `errors` from our application state:

    (state) => {
      return {
        errors: state.errors
      }
    }

And create a helper that dispatches our `signUp` asynchronous action:

    (dispatch) => {
      return {
        signUp: (email, password, password_confirm) => {
          return dispatch(signUp(email, password, password_confirm));
        }
      };
    }

Now that `JoinPage` is subscribed to changes to our store, we'll need to
move the logic that transforms our `errors` into a usable form from its
`constructor` into the `render` function, and replace the old `fetch`
logic with a call to `signUp`.

After trying to sign up with these changes, we'll see a runtime error in
our `UserMenu` component. It's expecting the newly signed-in user's
email address to be in `user.emails[0].address`. Changing this component
to pull the address from `user.email` fixes the errors.

The sign-up functionality is complete!