# Persisting Users
Unfortunately, if a user refreshes the page after signing up, they'll
lose their authenticated status. This means a user would have to sign-in
every time they load the application.

This issue is being caused by the fact that we're saving the `user` and
`jwt` objects exclusively in our in-memory application state. When we
reload the page, that state is reset.

Thankfully, we can fix this issue fairly quickly.

In our `signUp` thunk, once we recieve a successful response from the
server, we can store the `user` and `jwt` objects into local storage.

    localStorage.setItem("user", JSON.stringify(res.user));
    localStorage.setItem("jwt", res.jwt);

Similarly, when a user signs out we'll clear these local storage
entries:

    localStorage.removeItem("user");
    localStorage.removeItem("jwt");

Now we can popoulate our `initialState` with these `user` and `jwt`
values, if they exist in local storage:

    const user = localStorage.getItem("user");
    const jwt = localStorage.getItem("jwt");

    const initialState = {
      user: user ? JSON.parse(user) : user,
      jwt,
      ...

And now when a authenticated user refreshes the page, they'll stay
authenticated.