# List and Todo Models
We're getting to the point where we'll soon be wanting real data in our
application. To get start working with real data, we'll need to define
the schemas and models that describe that data.

Looking at our [`Lists`
schema](https://github.com/meteor/todos/blob/master/imports/api/lists/lists.js#L37-L42),
we know that each list needs a `name`, an `incompleteCount` (which we'll
call `incomplete_count`), and an optional reference to a user. We can
use Phoenix's `phoenix.gen.model` generator to create this model for us:

    mix phoenix.gen.model List lists name:string incomplete_count:integer user_id:references:users

Running this command creates a migration to create the `"users"` table
in our database. It also creates our `PhoenixTodos.List` model.

We can repeat this process for our `Todos` collection. Looking at the
[`Todos`
schema](https://github.com/meteor/todos/blob/master/imports/api/todos/todos.js#L39-L62),
we know we'll need a `text` field, a `checked` field, a reference to its
parent list, and a timestamp.

Once again, we can use the `phoenix.gen.model` generator to create this
model and migration for us:

    mix phoenix.gen.model Todo todos text:string checked:boolean list_id:references:lists

Nearly all of the code generated for us is perfect. We only need to make
one small tweak to our `PhoenixTodos.List` model. In addition to
specifying that it `belongs_to` the `PhoenixTodos.User` model, we need
to specify that each `PhoenixTodos.List` model `has_many`
`PhoenixTodos.Todo` children:

    has_many :todos, PhoenixTodos.Todo

Specifying this relationship on the parent `List` as well as the child
`Todo` model will most likely be very helpful down the line.