# Public Lists
Now that our database is populated with `Lists` and `Todos`, we're in a
position where we can start passing this data down the the client.

To keep things as similar to our original Meteor application as
possible, we'll be doing all of our commuication via WebSockets.
Specifically, we'll be using [Phoenix
Channels](http://www.phoenixframework.org/docs/channels).

We'll start by defining a `"lists.public"` channel. This channel will
emulate the `"lists.public"` publication in our Meteor application:

    channel "lists.public", PhoenixTodos.ListChannel

When a client joins this channel, we initially return all public lists:

    lists = List |> List.public |> Repo.all
    {:ok, lists, socket}

Where public lists are lists without an associated `User`:

    def public(query) do
      from list in query,
      where: is_nil(list.user_id)
    end

In order to send these lists down the wire, we need to use
[Poison](https://github.com/devinus/poison) to tell Phoenix how to
serialize our `List` objects:

    @derive {Poison.Encoder, only: [
      :id,
      :name,
      :incomplete_count,
      :user_id
    ]}

Now our client can connect to our server and join the `"lists.public"`
channel:

    socket.connect();
    socket.channel("lists.public", {})
      .join()

For each of the lists we receive back, well fire an `ADD_LIST` Redux
action. The resolver for this action simple pushes the `List` object
onto our application's `lists` array:

    return Object.assign({}, state, {
      lists: [...state.lists, action.list]
    });

And with that (and a few minor bug fixes), our application is now
showing lists pulled from the server!