# List Page
Now that are lists are being populated in the sidebar of our
application, we should pull in the components, layouts, and pages
necessary to render them.

We'll grab the `ListPageContainer`, `ListPage`, `ListHeader`, and
`TodoItem` React components from our original Meteor application and
move them into our Phoenix application.

So far, the main changes we've made to these components is to rename
variables to match our Ecto models (`incompleteCount` to
`incomplete_count`, and `userId` to `user_id`), and refactoring how we
fetch lists.

Instead of using Meteor collection to fetch lists from Minimongo, we
pull them directly out of our application's state:

    let id = props.params.id;
    let list = _.find(state.lists, list => list.id == id);

Lastly, we need to add our new `ListPageContainer` to our router:

    <Route path="lists/:id" component={ListPageContainer}/>

Now clicking on a list in our sidebar shows the (empty) list in the main
panel!