# Creating Public Lists
Now that we're loading and displaying all public lists, we'll want to be
able to create new lists.

To emulate our Meteor application, new lists will be given unique names,
starting with `"List A"`. If `"List A"` already exists, we'll use `"List
B"`, and so on.

We'll implement this functionality in a `create` function on our `List`
model. When `create` is called with no arguments, we'll default to using
`"List"` as the base name, and `"A"` as the suffix:

    def create, do: create("List", "A")

When `create` is called with two arguments (`name`, and `suffix`), we'll
find any lists that exist with the given name:

    PhoenixTodos.List
    |> findByName("#{name} #{suffix}")
    |> Repo.all

If we find one or more lists with that name, we'll increment our
`suffix` to the next character and try again:

    def handle_create_find(_, name, suffix) do
      [char] = to_char_list suffix
      create(name, to_string [char + 1])
    end

Otherwise, we know that our `name`/`suffix` combination is unique, so
we'll insert the list.

We'll call our `create` function whenever our `"lists.public"` channel
receives a `"create_list"` message:

    def handle_in("create_list", _, socket) do
      list = List.create

Once the list is added, we'll want to broadcast this change to all
connected clients, so they can add the new list to their UI:

    broadcast! socket, "add_list", list

Finally, we'll reply with the newly created list.

On the client, we removed all references to `insert.call` in favor of a
newly created `createList` action creator function. `createList` pushes
the `"create_list"` message up to the server, and handles all responses.