# List Ordering
Our last commit has a small issue. When lists were added, they appeared
at the end of the list, as expected. However, when we reloaded the
application, lists would appear in a seemingly random order.

To fix this, we need to order the lists by when they were inserted into
the database.

A quick way to do this is to add an `order_by` clause to our
`List.public` query:

    order_by: list.inserted_at,

Now our lists will be consistently ordered, even through refreshes.