# Adding Tasks
Now that we can add lists, we should be able to add tasks to our
lists.

We'll start by adding an `add_task` function to our `List` module.
`add_task` takes in the list's `id` that we're updating, and the `text`
of the new task.

After we grab our `list` from the database, we can use
`Ecto.build_assoc` to create a new `Task` associated with it:

    Ecto.build_assoc(list, :todos, text: text)
    |> Repo.insert!

Next, we'll need to increment the list's `incomplete_count`:

    list
    |> PhoenixTodos.List.changeset(%{
      incomplete_count: list.incomplete_count + 1
    })
    |> Repo.update!

Now we'll wire our new `add_task` model function up to a `"add_task"`
message handler on our `ListChannel`:

    def handle_in("add_task", %{
      "list_id" => list_id,
      "text" => text
    }, socket) do
      list = List.add_task(list_id, text)

Once we've added the list, we need to inform all subscribed clients that
the list has been updated. We'll do this by broadcasting a
`"update_list"` message:

    broadcast! socket, "update_list", list

Finally, we can replace our call to `insert.call` with a Redux thunk
that triggers our `"add_task"` channel message:

    this.props.addTask(this.props.list.id, input.value);

Now we can add new tasks to each of our todos!