# Checking Tasks
Next up on our feature list is giving users the ability to toggle tasks
as completed or incomplete.

To do this, we'll create a `set_checked_status` helper method in our
`List` model. Oddly, `set_checked_status` takes in a `todo_id` and a
`checked` boolean. This will likely be a good place for a future
refacotr...

The `set_checked_status` function starts by grabbing the specified todo
and it's associated todo:

    todo = Repo.get(PhoenixTodos.Todo, todo_id)
    |> Repo.preload(:list)
    list = todo.list

Next, it uses `checked` to determine if we'll be incrementing or
decrementing `incomplete_count` on our list:

    inc = if (checked), do: - 1, else: 1

We can update our todo by setting the `checked` field:

    todo
    |> PhoenixTodos.Todo.changeset(%{
      checked: checked
    })
    |> Repo.update!

And we can update our list by setting the `incomplete_count` field:

    list
    |> PhoenixTodos.List.changeset(%{
      incomplete_count: list.incomplete_count + inc
    })
    |> Repo.update!

Now that we have a functional helper method in our model, we can call it
whenever we receive a `"set_checked_status"` message in our list
channel:

    def handle_in("set_checked_status", %{
        "todo_id" => todo_id,
        "status" => status
      }, socket) do
      list = List.set_checked_status(todo_id, status)

Lastly, we'll broadcast a `"update_list"` message to all connected
clients so they can see this change in realtime.

Now we can replace our call to the old `setCheckedStatus` Meteor method
with a call to an asynchronous action creator which pushes our
`"set_checked_status"` message up to the server.

With that, our users can check and uncheck todos.