# Sorting Tasks
Just like our lists, our tasks are having a sorting problem. Checking
and unchecking a task will randomize its position in the list.

There are two ways to solve this issue. We can either sort our `todos`
when we `Preload` them on the server, or we can do our sorting on the
client. Let's go with the second option.

Let's sort primarily based on the "created at" timestamp of each task.
To do this, we'll need to serialize the `inserted_at` timestamp for each
task we send to the client.

We can then sort our `todos` on this timestamp before sending our
`<TodoItem>` components:

    .sort((a, b) => {
      return new Date(a.inserted_at) - new Date(b.inserted_at);
    })

We'll also want to have a secondary sort on the task's text to break and
ties that may occur (especially in seed data):

    .sort((a, b) => {
      let diff = new Date(a.inserted_at) - new Date(b.inserted_at);
      return diff == 0 ? a.text > b.text : diff;
    })

With those changes, our tasks order themselves correctly in each list.