# Update List Name
The next piece of functionality we need to knock out is the ability to
rename lists.

To do this, we'll create a helper method on our `List` model called
`update_name`. This method simply changes the `name` of the given
`List`:

    Repo.get(PhoenixTodos.List, id)
    |> changeset(%{
      name: name
    })
    |> Repo.update!

We'll create a new channel event, `"update_name"`, to handle name change
requests, and we'll wire up a Redux thunk to push a `"update_name"`
event onto our channel:

    channel.push("update_name", { list_id, name })
      .receive("ok", (list) => {
        dispatch(updateNameSuccess());
      })
      .receive("error", () => dispatch(updateNameFailure()))
      .receive("timeout", () => dispatch(updateNameFailure()));

After wiring up the rest of the necessary Redux plumbing, we're able to
update the names of our lists.