# Authenticated Sockets
Now that we've implemented the bulk of the unauthenticated functionality
in our application, we need to turn our attention to authenticated
functionality.

To do that, we'll need to authenticate the channel we're using to
communicate with the client. We can do this with a custom `connect`
function that verified the user's provided `guardian_token`:

    def connect(%{"guardian_token" => jwt}, socket) do
      case sign_in(socket, jwt) do
        {:ok, authed_socket, guardian_params} ->
          {:ok, authed_socket}
        _ ->
          {:ok, socket}
      end
    end

If the user is authenticated correctly, we'll swap their socket out for
an `auth_socket`, which can be used to access the current user's object
and claims.

If the user doesn't provide a `guardian_token`, we'll fall back to our
old `connect` function:

  def connect(_params, socket) do
    {:ok, socket}
  end

All of the old functionality will continue to work as expected.