# Connect Socket Thunk
That that our entire socket connection will be authenticated or
unauthenticated, we need to prepare ourselves to re-establish the
connection every time we log in/out.

To start, we'll need to clear our local set of `lists` every time we
connect to the socket.

We also need to trigger a call to our `joinListsChannel` thunk every
time we connect.