# Make Public
Just as we let users make their lists private, we need to let them make
their private lists public again.

We'll do this by adding a `"make_public"` channel event that sets the
`user_id` field on the specified list to `nil` and broadcasts an
`"update_list"` event.

    list = List.make_public(list_id)
    |> Repo.preload(:todos)

    broadcast! socket, "update_list", list

Unfortunately, this introduces a situation where lists are added back
into the UI through a `"update_list"` event rather than a `"add_list"`
event.

To handle this, we need to check if the `"UPDATE_LIST"` Redux reducer
actually found the list it was trying to update. If it didn't, we'll
push the list to the end of the list, adding it to the UI:

    if (!found) {
      lists.push(action.list);
    }

And with that, users can make their private lists public.